api-server/requirements.txt
connexion == 1.1.15
flask-cors >= 3.0.3
pymongo>=3.5.1,<3.6
python_dateutil == 2.6.0
setuptools >= 21.0.0
